crawler_app/scrape/requirements.txt
djangorestframework==3.15.2
psycopg2-binary
django==5.0.6
pytest==8.2.2
psycopg2==2.9.9
django-environ==0.4.3
requests==2.32.3